Refactoring Bakery Search Input 분리
- refactoring BakeryMap 에서 SearchInput 분리
- rename BakeryMap 이름을 index.tsx -> BakeryMap.tsx 로 수정
